move the first batch of unit tests into the proper job
